Complete step_4

+++ src/main/java/Color.java
@@ -1,18 +1,18 @@
 public class Color {
-    int[] e = new int[3];
-    public Color (int r, int g, int b) {
+    double[] e = new double[3];
+    public Color (double r, double g, double b) {
         e[0] = r;
         e[1] = g;
         e[2] = b;
     }
 
-    public int r() {return e[0];}
-    public int g() {return e[1];}
-    public int b() {return e[2];}
+    public double r() {return e[0];}
+    public double g() {return e[1];}
+    public double b() {return e[2];}
 
     public static Color fromVec3(Vec3 v) {
 
-        return new Color((int) v.x(), (int) v.y(), (int) v.z());
+        return new Color(v.x(), v.y(), v.z());
     }
 
     public String toString(Color pixelColor) {
